+++ presentation/src/main/java/com/crskdev/mealcalculator/presentation/common/entities/Mappers.kt
@@ -36,7 +36,7 @@ fun FoodVM.toDomain(): Food = Food(
         fat.unsaturated.toFloat()
     ),
     proteins.toFloat(),
-    gi.toInt()
+    gi.toFloat()
 )
 
 fun FoodVM.toDomainUnchecked(): FoodUnchecked =
